requirements.txt
streamlit
json
pandas
os
dateformat
google.generativeai
re